refactor: share system info detection logic
